Retornando o arquivo PessoaService

+++ CatalagoandoColecoes/src/main/java/br/edu/iff/ccc/bsi/webdev/service/PessoaService.java
@@ -7,7 +7,6 @@ import java.util.Calendar;
 import java.util.Locale;
 import java.util.Map;
 
-import org.hibernate.mapping.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -33,50 +32,31 @@ public class PessoaService {
 	@Autowired
 	private ColecaoService colecaoService = new ColecaoService();
 	
-	//public boolean save(Usuario user, Pessoa pessoa, Endereco endereco) {
 	public Pessoa save(Usuario user, Pessoa pessoa, Endereco endereco) {
-//		pessoa.setUsuario(usuarioService.save(user));
 		if(rep.consultaIdPessoa(pessoa.getCpf()) == null) {
-				pessoa.setUsuario(user);
-				pessoa.setEndereco(endereco);
-				Pessoa p = rep.save(pessoa);
-				return p;
+			pessoa.setUsuario(user);
+			pessoa.setEndereco(endereco);
+			Pessoa p = rep.save(pessoa);
+			return p;
 		} else {
 			System.out.println("Pessoa já está cadastrada!");
 		}
-		//rep.save(pessoa);
-		//return true;
+				
 		return null;
 	}
 	
 	
-//	public Pessoa remove(Usuario user, Pessoa pessoa, Endereco endereco) {
-//	public Pessoa remove(Pessoa pessoa) {
 	public Pessoa remove(String cpf) {
-//		pessoa.setUsuario(user);
-		//usuarioService.remove(Long.parseLong(String.valueOf(rep.consultaFKUsuario(pessoa.getCpf()))));
-//		System.out.println("FK_USUARIO: "+rep.consultaFKUsuario(pessoa.getCpf()));
-//		usuarioService.remove(rep.consultaFKUsuario(pessoa.getCpf()));
-//		pessoa.setEndereco(endereco);
-		//long id = Long.parseLong(rep.consultaIdPessoa(pessoa.getCpf()));
 		
-		
-		
-		//Map<String,String> pessoaP = rep.consultaPessoa(pessoa.getCpf());
-		//rep.deleteById(Long.parseLong(String.valueOf(pessoaP.get("ID"))));		
-
-		long id = Long.parseLong(String.valueOf(rep.consultaIdPessoa(cpf)));
-		Pessoa pessoa = this.consultaPessoa(cpf);
-//		long idUsuario = rep.consultaFKUsuario(pessoa.getCpf());
-		rep.deleteById(id);
-		
-//		usuarioService.remove(idUsuario);
-		if(pessoa.equals(null)) {
-			//return false;
-			return null;
-		}
-		//rep.save(pessoa);
-		return pessoa;
+		if(rep.consultaIdPessoa(cpf) != null) {
+			long id = Long.parseLong(String.valueOf(rep.consultaIdPessoa(cpf)));
+			Pessoa pessoa = this.consultaPessoa(cpf);
+			rep.deleteById(id);			
+			return pessoa;
+		} else {
+			System.out.println("Pessoa não está cadastrada!");
+		}	
+		return null;
 	}
 	
 	
@@ -103,7 +83,7 @@ public class PessoaService {
 			Long idUsuario = this.consultaIDUsuario(cpf);
 			usuario = usuarioService.consultaUsuario(idUsuario);
 			pessoa.setUsuario(usuario);			
-			
+					
 			if(colecaoService.consultaColecao(cpf) != null) {
 				Colecao c = colecaoService.consultaColecao(cpf);
 				pessoa.setColecao(c);
@@ -116,7 +96,6 @@ public class PessoaService {
 		}
 	}
 	
-	
 	public ArrayList<Pessoa> consultaPessoas() {
 		return (ArrayList<Pessoa>) rep.findAll();
 	}
@@ -126,44 +105,23 @@ public class PessoaService {
 		return Long.parseLong(String.valueOf(pessoaConsultada.get("FK_USUARIO")));
 	}
 	
-	//public Pessoa atualizar(Usuario user, Pessoa pessoa, Endereco endereco) {
+	
+	
 	public Pessoa atualizar(Usuario user, Pessoa pessoa, Endereco endereco, String cpf) {
-//		pessoa.setUsuario(user);
-		
-//		boolean testeUsuario, testePessoa;
-//		long idUsuario = rep.consultaFKUsuario(pessoa.getCpf());
-//		user.setID(idUsuario);
-//		testeUsuario = usuarioService.atualizar(user);
-//		pessoa.setUsuario(user);
-//		pessoa.setEndereco(endereco);
-//		testePessoa = (rep.atualizar(pessoa.getCpf(), pessoa.getNome(), pessoa.getEmail(), pessoa.getEndereco().getCEP(), pessoa.getEndereco().getRua(), pessoa.getEndereco().getNumero(), pessoa.getEndereco().getBairro(), pessoa.getEndereco().getCidade(), pessoa.getEndereco().getEstado()) == 1);
-//		if((testeUsuario)&&(testePessoa)) {
-//			return pessoa;
-//		} else {
-//			return null;
-//		}
-		
-		
-		long id = Long.parseLong(String.valueOf(rep.consultaIdPessoa(cpf)));
-		pessoa.setID(id);
-		user.setID(this.consultaIDUsuario(cpf));
-		pessoa.setUsuario(user);
-		pessoa.setEndereco(endereco);
-		return rep.saveAndFlush(pessoa);
-			
-//		pessoa.setUsuario(usuarioService.atualizar(user));
-//		pessoa.setEndereco(endereco);
-//		int i = rep.atualizar(pessoa.getCpf(), pessoa.getNome(), pessoa.getEmail(), pessoa.getEndereco().getCEP(), pessoa.getEndereco().getRua(), pessoa.getEndereco().getNumero(), pessoa.getEndereco().getBairro(), pessoa.getEndereco().getCidade(), pessoa.getEndereco().getEstado());
-//		System.out.println("i: "+i);
-
-//		return pessoa;
-		
-		//erator<Pessoa> pessoas = (Iterator<Pessoa>) rep.findAll();
-		
-//		Pessoa p = rep.save(pessoa);
-		//rep.flush();
-//		return p;
+		if(rep.consultaIdPessoa(cpf) != null) {
+	//		long id = Long.parseLong(String.valueOf(rep.consultaIdPessoa(pessoa.getCpf())));
+			Long id = rep.consultaIdPessoa(cpf);
+			pessoa.setID(id);
+			user.setID(this.consultaIDUsuario(cpf));
+			pessoa.setUsuario(user);
+			pessoa.setEndereco(endereco);
+			return rep.saveAndFlush(pessoa);
+		} else {
+			System.out.println("Pessoa não está cadastrada!");
+		}
+		 return null;
 	}
+	
 	
 	public ArrayList<Pessoa> consultaPessoasPorNome(String nome) {
 		String nomeP = "%"+nome+"%";
@@ -179,7 +137,6 @@ public class PessoaService {
 	
 	
 	public Pessoa criarColecao(Pessoa pessoa, Item item, Map<String,String> dadosColecao) {
-//		Long idPessoa = rep.consultaIdPessoa(pessoa.getCpf());
 		Colecao colecao = new Colecao();
 		colecao.setNome(dadosColecao.get("nome"));
 		colecao.setObservacao(dadosColecao.get("observacao"));
@@ -203,12 +160,10 @@ public class PessoaService {
 		if(rep.consultaFKColecao(pessoa.getCpf()) == null) {
 			if(rep.consultaIdPessoa(pessoa.getCpf()) != null) {
 				if(itemService.consultaIdItem(item.getIsbn()) != null) {
-		//			pessoa.setID(idPessoa);
 					Pessoa p = this.consultaPessoa(pessoa.getCpf());
 					item.setID(idItem);
 					colecao.addItem(item);
 					p.setColecao(colecao);			
-		//			return colecaoService.save(pessoa, colecao, item);
 					return rep.saveAndFlush(p);
 				} else {
 					System.out.println("Item não está cadastrado!");
@@ -261,6 +216,7 @@ public class PessoaService {
 		return null;
 	}
 	
+	
 	public Pessoa removerColecao(String cpf) {
 		String cpfPessoa = cpf;
 		if(rep.consultaIdPessoa(cpfPessoa) != null) {
@@ -280,7 +236,6 @@ public class PessoaService {
 		
 		return null;
 	}
-	
 	
 	
 	public Pessoa addItem(String cpf, String isbn) {
@@ -366,7 +321,6 @@ public class PessoaService {
 		}
 		return null;
 	}
-	
 	
 	public ArrayList<Colecao> consultaColecoes() {
 		return colecaoService.consultaColecoes();
